+++ src/main/spi/spi.h
@@ -3,7 +3,7 @@
 #include <stdint.h>
 #include <stddef.h>
 
-int configure_pins();
+uint8_t configure_pins();
 void reset_pins();
 uint8_t spi_read_write(uint8_t* buff, size_t buff_size);
